Rename utils/utils.js to utils/index.js
